adminLogin/requirements.txt
pyJwt
bcrypt
pymongo